YQL-20229: [refactoring] split nodes in separate files
Zero diff, поразносил по файлам
commit_hash:f668f5b8e6b40f977a15ea1916ff044972459f15
