+++ yql/essentials/minikql/comp_nodes/mkql_factory.cpp
@@ -37,6 +37,7 @@
 #include "mkql_ensure.h"
 #include "mkql_enumerate.h"
 #include "mkql_exists.h"
+#include "mkql_expand_map.h"
 #include "mkql_extend.h"
 #include "mkql_filter.h"
 #include "mkql_flatmap.h"
@@ -69,6 +70,7 @@
 #include "mkql_map_join.h"
 #include "mkql_match_recognize.h"
 #include "mkql_multimap.h"
+#include "mkql_narrow_map.h"
 #include "mkql_next_value.h"
 #include "mkql_nop.h"
 #include "mkql_now.h"
