+++ yql/essentials/minikql/comp_nodes/mkql_narrow_map.h
@@ -1,8 +1,9 @@
 #pragma once
+
 #include <yql/essentials/minikql/computation/mkql_computation_node.h>
 
 namespace NKikimr::NMiniKQL {
 
-IComputationNode* WrapWideMap(TCallable& callable, const TComputationNodeFactoryContext& ctx);
+IComputationNode* WrapNarrowMap(TCallable& callable, const TComputationNodeFactoryContext& ctx);
 
 }
