+++ yql/essentials/minikql/comp_nodes/ya.make.inc
@@ -47,6 +47,7 @@ SET(ORIG_SOURCES
     mkql_ensure.cpp
     mkql_enumerate.cpp
     mkql_exists.cpp
+    mkql_expand_map.cpp
     mkql_extend.cpp
     mkql_factory.cpp
     mkql_filter.cpp
@@ -85,6 +86,7 @@ SET(ORIG_SOURCES
     mkql_match_recognize_rows_formatter.cpp
     mkql_multihopping.cpp
     mkql_multimap.cpp
+    mkql_narrow_map.cpp
     mkql_next_value.cpp
     mkql_nop.cpp
     mkql_now.cpp
